Pseudobam now has single end mode working, all flags are implemented
